chore: moved functions into clearer files
