+++ clwdevops/connect.py
@@ -1,8 +1,10 @@
 import re
-import subprocess
 from dataclasses import dataclass, field
+from pathlib import PosixPath
+from datetime import datetime as dt
+import logging
 
-from omegaconf import DictConfig, OmegaConf
+log = logging.getLogger(__name__)
 
 
 @dataclass
@@ -27,7 +29,41 @@ class UploadId:
         return cls(sid)
 
 
-def load_sops_file(fname: str) -> DictConfig:
-    sops = subprocess.run(["sops", "-d", fname], capture_output=True, text=True).stdout
-    cfg = OmegaConf.create(sops)
-    return cfg
+def convert_filename_to_datetime(filename: PosixPath) -> dt:
+    """Convert filename to datetime
+    Assumes filename is in format path/to/file/MMDDYY_XX_HHMMSS.ext"""
+
+    if not isinstance(filename, PosixPath):
+        raise TypeError("filename must be a PosixPath")
+
+    basename = filename.stem  # Get the filename without the extension
+    components = basename.split("_")  # Split the filename into its components
+
+    date = components[0]  # Get the date
+    time = components[2]  # Get the time
+
+    try:
+        timestamp = dt.strptime(
+            f"{date} {time}", "%m%d%y %H%M%S"
+        )  # Combine the date and time into a timestamp
+    except ValueError:
+        log.error(f"Invalid filename format {filename}")
+        return
+
+    return timestamp
+
+
+def timestamps_to_study_times(timestamps: list[dt]) -> dict:
+
+    if not isinstance(timestamps, list):
+        raise TypeError("timestamps must be a list")
+
+    studyTimeLabels = ["Timestamp", "StudyDay", "Hour", "HourOfDay"]
+
+    studyTime = {k: [] for k in studyTimeLabels}
+    studyTime["Timestamp"] = timestamps
+    studyTime["StudyDay"] = [i // 24 + 1 for i in range(len(timestamps))]
+    studyTime["Hour"] = [i + 1 for i in range(len(timestamps))]
+    studyTime["HourOfDay"] = [(x - 1) % 24 + 1 for x in studyTime["Hour"]]
+
+    return studyTime
